YAuurrrr we got the github network working

+++ Connection.java
@@ -1,9 +1,9 @@
 public class Connection {
     private User user1;
     private User user2;
-    private int weight; // Strength of the relationship
+    private double weight; // Strength of the relationship
 
-    public Connection(User user1, User user2, int weight) {
+    public Connection(User user1, User user2, double weight) {
         this.user1 = user1;
         this.user2 = user2;
         this.weight = weight;
@@ -17,11 +17,11 @@ public class Connection {
         return user2;
     }
 
-    public int getWeight() {
+    public double getWeight() {
         return weight;
     }
 
-    public void setWeight(int weight) {
+    public void setWeight(double weight) {
         this.weight = weight;
     }
 }
